hoisted out constants and moved color utils to module
