+++ src/extras/gizmo/constants.js
@@ -1,0 +1,71 @@
+/**
+ * Local coordinate space.
+ *
+ * @type {string}
+ * @category Gizmo
+ */
+export const GIZMOCOORD_LOCAL = 'local';
+
+/**
+ * World coordinate space.
+ *
+ * @type {string}
+ * @category Gizmo
+ */
+export const GIZMOCOORD_WORLD = 'world';
+
+/**
+ * Gizmo axis for the line X.
+ *
+ * @type {string}
+ */
+export const GIZMOAXIS_X = 'x';
+
+/**
+ * Gizmo axis for the line Y.
+ *
+ * @type {string}
+ */
+export const GIZMOAXIS_Y = 'y';
+
+/**
+ * Gizmo axis for the line Z.
+ *
+ * @type {string}
+ */
+export const GIZMOAXIS_Z = 'z';
+
+/**
+ * Gizmo axis for the plane YZ.
+ *
+ * @type {string}
+ */
+export const GIZMOAXIS_YZ = 'yz';
+
+/**
+ * Gizmo axis for the plane XZ.
+ *
+ * @type {string}
+ */
+export const GIZMOAXIS_XZ = 'xz';
+
+/**
+ * Gizmo axis for the plane XY.
+ *
+ * @type {string}
+ */
+export const GIZMOAXIS_XY = 'xy';
+
+/**
+ * Gizmo axis for all directions XYZ.
+ *
+ * @type {string}
+ */
+export const GIZMOAXIS_XYZ = 'xyz';
+
+/**
+ * Gizmo axis for facing the camera (F).
+ *
+ * @type {string}
+ */
+export const GIZMOAXIS_F = 'f';
