+++ src/extras/gizmo/rotate-gizmo.js
@@ -4,9 +4,10 @@ import { Quat } from '../../core/math/quat.js';
 import { Mat4 } from '../../core/math/mat4.js';
 import { Vec3 } from '../../core/math/vec3.js';
 import { PROJECTION_ORTHOGRAPHIC, PROJECTION_PERSPECTIVE } from '../../scene/constants.js';
+
 import { AxisDisk } from './axis-shapes.js';
-import { GIZMOCOORD_LOCAL } from './gizmo.js';
-import { GIZMOAXIS_F, GIZMOAXIS_X, GIZMOAXIS_Y, GIZMOAXIS_Z, TransformGizmo } from "./transform-gizmo.js";
+import { GIZMOCOORD_LOCAL, GIZMOAXIS_F, GIZMOAXIS_X, GIZMOAXIS_Y, GIZMOAXIS_Z } from './constants.js';
+import { TransformGizmo } from './transform-gizmo.js';
 
 /**
  * @import { AppBase } from '../../framework/app-base.js'
