+++ src/extras/gizmo/scale-gizmo.js
@@ -1,8 +1,9 @@
 import { Vec3 } from '../../core/math/vec3.js';
 import { Quat } from '../../core/math/quat.js';
+
 import { AxisBoxCenter, AxisBoxLine, AxisPlane } from './axis-shapes.js';
-import { GIZMOCOORD_LOCAL } from './gizmo.js';
-import { GIZMOAXIS_X, GIZMOAXIS_XYZ, GIZMOAXIS_Y, GIZMOAXIS_Z, TransformGizmo } from './transform-gizmo.js';
+import { GIZMOCOORD_LOCAL, GIZMOAXIS_X, GIZMOAXIS_XYZ, GIZMOAXIS_Y, GIZMOAXIS_Z } from './constants.js';
+import { TransformGizmo } from './transform-gizmo.js';
 
 /**
  * @import { AppBase } from '../../framework/app-base.js'
